requirements.txt
pandas
numpy
scikit-learn
joblib
shap
matplotlib
seaborn
fpdf
google-generativeai
python-dotenv